Add pipeline support
This commit also significantly changes the behavior of the plugin.

When a node is set to leniently shut down, it will only permit any jobs currently executing on it to finish.
In the case of pipeline jobs, that includes all of their PlaceholderTasks.

When Jenkins is set to leniently shut down, WorkflowRuns are treated the same as other builds.
PlaceholderTasks are always allowed to execute, as they must have been spawned from a run that was already permitted to begin.

When storing queue IDs for comparison, we only store the ID of the WorkflowRun - never the PlaceholderTasks